fix: correção dos .stories
