+++ src/components/admin/browser-select/__mocks__/BrowserSelect.ts
@@ -12,7 +12,7 @@ const data = [
 ]
 
 const service = {
-	get(params) {
+	get(params: any): Promise<any> {
 		console.log(params)
 		const meta = {
 			current_page: params.page,
